+++ egui_web/src/webgl1.rs
@@ -299,7 +299,7 @@ impl crate::Painter for WebGlPainter {
             }
         }
 
-        crate::console_error("Failed to query max texture size");
+        tracing::error!("Failed to query max texture size");
 
         2048
     }